+++ test/e2e/index.ts
@@ -14,11 +14,12 @@ import {
   AppSignal,
   AppWebsocket,
   CallZomeRequest,
+  CreateCloneCellRequest,
 } from "../../src/api/app/index.js";
 import { WsClient } from "../../src/api/client.js";
+import { CloneId } from "../../src/api/common.js";
 import {
-  CONFIG_PATH,
-  CONFIG_PATH_1,
+  cleanSandboxConductors,
   FIXTURE_PATH,
   installAppAndDna,
   launch,
@@ -55,6 +56,7 @@ test(
 
     const path = `${FIXTURE_PATH}/test.dna`;
     const hash = await admin.registerDna({
+      modifiers: {},
       path,
     });
 
@@ -157,7 +159,9 @@ test(
     // install from hash and network seed
     const newHash = await admin.registerDna({
       hash,
-      network_seed: "123456",
+      modifiers: {
+        network_seed: "123456",
+      },
     });
     t.ok(newHash);
 
@@ -247,6 +251,7 @@ test(
 
     const dnaBundle: DnaBundle = decode(encodedDnaBundle.buffer) as DnaBundle;
     const hash = await admin.registerDna({
+      modifiers: {},
       bundle: dnaBundle,
     });
     t.ok(hash);
@@ -377,7 +382,6 @@ test(
 test(
   "can handle canceled response",
   withConductor(ADMIN_PORT, async (t: Test) => {
-    // const client = await WsClient.connect(`ws://localhost:${ADMIN_PORT}`);A
     const client = new WsClient({
       send: () => {
         /* do nothing */
@@ -477,94 +481,94 @@ test("error is catchable when holochain socket is unavailable", async (t: Test) 
 });
 
 test("can inject agents", async (t: Test) => {
-  const [conductor1, l1] = await launch(ADMIN_PORT, CONFIG_PATH);
-  const [conductor2, l2] = await launch(ADMIN_PORT_1, CONFIG_PATH_1);
-  try {
-    const installed_app_id = "app";
-    const admin1 = await AdminWebsocket.connect(`ws://localhost:${ADMIN_PORT}`);
-    const admin2 = await AdminWebsocket.connect(
-      `ws://localhost:${ADMIN_PORT_1}`
-    );
-    const agent_key_1 = await admin1.generateAgentPubKey();
-    t.ok(agent_key_1);
-    const agent_key_2 = await admin2.generateAgentPubKey();
-    t.ok(agent_key_2);
-    const role = "thedna";
-    const path = `${FIXTURE_PATH}/test.dna`;
-    const hash = await admin1.registerDna({ path });
-    t.ok(hash);
-    let result = await admin1.installApp({
-      installed_app_id,
-      agent_key: agent_key_1,
-      dnas: [{ hash, role_id: role }],
-    });
-    t.ok(result);
-    const app1_cell = result.cell_data[0].cell_id;
-    const activeApp1Info = await admin1.enableApp({ installed_app_id }, 1000);
-    t.deepEqual(activeApp1Info.app.status, { running: null });
-    t.equal(activeApp1Info.app.cell_data[0].role_id, role);
-    t.equal(activeApp1Info.app.installed_app_id, installed_app_id);
-    t.equal(activeApp1Info.errors.length, 0);
+  const conductor1 = await launch(ADMIN_PORT);
+  const conductor2 = await launch(ADMIN_PORT_1);
+  const installed_app_id = "app";
+  const admin1 = await AdminWebsocket.connect(`ws://localhost:${ADMIN_PORT}`);
+  const admin2 = await AdminWebsocket.connect(`ws://localhost:${ADMIN_PORT_1}`);
+  const agent_key_1 = await admin1.generateAgentPubKey();
+  t.ok(agent_key_1);
+  const agent_key_2 = await admin2.generateAgentPubKey();
+  t.ok(agent_key_2);
+  const role = "thedna";
+  const path = `${FIXTURE_PATH}/test.dna`;
+  const hash = await admin1.registerDna({ path, modifiers: {} });
+  t.ok(hash);
+  let result = await admin1.installApp({
+    installed_app_id,
+    agent_key: agent_key_1,
+    dnas: [{ hash, role_id: role }],
+  });
+  t.ok(result);
+  const app1_cell = result.cell_data[0].cell_id;
+  const activeApp1Info = await admin1.enableApp({ installed_app_id }, 1000);
+  t.deepEqual(activeApp1Info.app.status, { running: null });
+  t.equal(activeApp1Info.app.cell_data[0].role_id, role);
+  t.equal(activeApp1Info.app.installed_app_id, installed_app_id);
+  t.equal(activeApp1Info.errors.length, 0);
 
-    await delay(500);
+  await delay(500);
 
-    // after activating an app requestAgentInfo should return the agentid
-    // requesting info with null cell_id should return all agents known about.
-    // otherwise it's just agents know about for that cell
-    const conductor1_agentInfo = await admin1.requestAgentInfo({
-      cell_id: null,
-    });
-    t.equal(conductor1_agentInfo.length, 1);
+  // after activating an app requestAgentInfo should return the agentid
+  // requesting info with null cell_id should return all agents known about.
+  // otherwise it's just agents know about for that cell
+  const conductor1_agentInfo = await admin1.requestAgentInfo({
+    cell_id: null,
+  });
+  t.equal(conductor1_agentInfo.length, 1);
 
-    // agent2 with no activated apps there are no agents
-    let conductor2_agentInfo = await admin2.requestAgentInfo({ cell_id: null });
-    t.equal(conductor2_agentInfo.length, 0);
+  // agent2 with no activated apps there are no agents
+  let conductor2_agentInfo = await admin2.requestAgentInfo({ cell_id: null });
+  t.equal(conductor2_agentInfo.length, 0);
 
-    // but, after explicitly injecting an agent, we should see it
-    await admin2.addAgentInfo({ agent_infos: conductor1_agentInfo });
-    conductor2_agentInfo = await admin2.requestAgentInfo({ cell_id: null });
-    t.equal(conductor2_agentInfo.length, 1);
-    t.deepEqual(conductor1_agentInfo, conductor2_agentInfo);
+  // but, after explicitly injecting an agent, we should see it
+  await admin2.addAgentInfo({ agent_infos: conductor1_agentInfo });
+  conductor2_agentInfo = await admin2.requestAgentInfo({ cell_id: null });
+  t.equal(conductor2_agentInfo.length, 1);
+  t.deepEqual(conductor1_agentInfo, conductor2_agentInfo);
 
-    // now install the app and activate it on agent 2.
-    await admin2.registerDna({
-      path,
-    });
-    t.ok(hash);
-    result = await admin2.installApp({
-      installed_app_id,
-      agent_key: agent_key_2,
-      dnas: [{ hash, role_id: role }],
-    });
-    t.ok(result);
-    const app2_cell = result.cell_data[0].cell_id;
-    const activeApp2Info = await admin2.enableApp({ installed_app_id });
-    t.deepEqual(activeApp2Info.app.status, { running: null });
-    t.equal(activeApp2Info.app.cell_data[0].role_id, role);
-    t.equal(activeApp2Info.app.installed_app_id, installed_app_id);
-    t.equal(activeApp2Info.errors.length, 0);
+  // now install the app and activate it on agent 2.
+  await admin2.registerDna({
+    modifiers: {},
+    path,
+  });
+  t.ok(hash);
+  result = await admin2.installApp({
+    installed_app_id,
+    agent_key: agent_key_2,
+    dnas: [{ hash, role_id: role }],
+  });
+  t.ok(result);
+  const app2_cell = result.cell_data[0].cell_id;
+  const activeApp2Info = await admin2.enableApp({ installed_app_id });
+  t.deepEqual(activeApp2Info.app.status, { running: null });
+  t.equal(activeApp2Info.app.cell_data[0].role_id, role);
+  t.equal(activeApp2Info.app.installed_app_id, installed_app_id);
+  t.equal(activeApp2Info.errors.length, 0);
 
-    await delay(500);
-    // observe 2 agent infos
-    conductor2_agentInfo = await admin2.requestAgentInfo({ cell_id: null });
-    t.equal(conductor2_agentInfo.length, 2);
+  await delay(500);
+  // observe 2 agent infos
+  conductor2_agentInfo = await admin2.requestAgentInfo({ cell_id: null });
+  t.equal(conductor2_agentInfo.length, 2);
 
-    // now confirm that we can ask for just one cell
-    await admin1.addAgentInfo({ agent_infos: conductor2_agentInfo });
-    const app1_agentInfo = await admin1.requestAgentInfo({
-      cell_id: app1_cell,
-    });
-    t.equal(app1_agentInfo.length, 1);
-    const app2_agentInfo = await admin2.requestAgentInfo({
-      cell_id: app2_cell,
-    });
-    t.equal(app2_agentInfo.length, 1);
-  } finally {
-    conductor1.kill();
-    conductor2.kill();
-    l1.kill();
-    l2.kill();
+  // now confirm that we can ask for just one cell
+  await admin1.addAgentInfo({ agent_infos: conductor2_agentInfo });
+  const app1_agentInfo = await admin1.requestAgentInfo({
+    cell_id: app1_cell,
+  });
+  t.equal(app1_agentInfo.length, 1);
+  const app2_agentInfo = await admin2.requestAgentInfo({
+    cell_id: app2_cell,
+  });
+  t.equal(app2_agentInfo.length, 1);
+
+  if (conductor1.pid) {
+    process.kill(-conductor1.pid);
   }
+  if (conductor2.pid) {
+    process.kill(-conductor2.pid);
+  }
+  await cleanSandboxConductors();
 });
 
 test(
@@ -636,5 +640,180 @@ test(
       installedApp1.cell_data[0].cell_id[0],
       installedApp2.cell_data[0].cell_id[0]
     );
+  })
+);
+
+test(
+  "can create a callable clone cell",
+  withConductor(ADMIN_PORT, async (t: Test) => {
+    const { installed_app_id, role_id, client } = await installAppAndDna(
+      ADMIN_PORT
+    );
+    const info = await client.appInfo({ installed_app_id });
+
+    const createCloneCellParams: CreateCloneCellRequest = {
+      app_id: installed_app_id,
+      role_id,
+      modifiers: {
+        network_seed: "clone-0",
+      },
+    };
+    const cloneCell = await client.createCloneCell(createCloneCellParams);
+
+    const expectedCloneId = new CloneId(role_id, 0).toString();
+    t.equal(cloneCell.role_id, expectedCloneId, "correct clone id");
+    t.deepEqual(
+      cloneCell.cell_id[1],
+      info.cell_data[0].cell_id[1],
+      "clone cell agent key matches base cell agent key"
+    );
+    const params: CallZomeRequest = {
+      cap_secret: null,
+      cell_id: cloneCell.cell_id,
+      zome_name: TEST_ZOME_NAME,
+      fn_name: "foo",
+      provenance: fakeAgentPubKey(),
+      payload: null,
+    };
+    const response = await client.callZome(params);
+    t.equal(
+      response,
+      "foo",
+      "clone cell can be called with same zome call as base cell"
+    );
+  })
+);
+
+test(
+  "can archive a clone cell",
+  withConductor(ADMIN_PORT, async (t: Test) => {
+    const { installed_app_id, role_id, client } = await installAppAndDna(
+      ADMIN_PORT
+    );
+    const createCloneCellParams: CreateCloneCellRequest = {
+      app_id: installed_app_id,
+      role_id,
+      modifiers: {
+        network_seed: "clone-0",
+      },
+    };
+    const cloneCell = await client.createCloneCell(createCloneCellParams);
+
+    await client.archiveCloneCell({
+      app_id: installed_app_id,
+      clone_cell_id: cloneCell.cell_id,
+    });
+
+    const appInfo = await client.appInfo({ installed_app_id });
+    t.equal(
+      appInfo.cell_data.length,
+      1,
+      "archived clone cell is not part of app info"
+    );
+    const params: CallZomeRequest = {
+      cap_secret: null,
+      cell_id: cloneCell.cell_id,
+      zome_name: TEST_ZOME_NAME,
+      fn_name: "foo",
+      provenance: fakeAgentPubKey(),
+      payload: null,
+    };
+    try {
+      await client.callZome(params);
+      t.fail();
+    } catch (error) {
+      t.pass("archived clone call cannot be called");
+    }
+  })
+);
+
+test(
+  "can restore an archived clone cell",
+  withConductor(ADMIN_PORT, async (t: Test) => {
+    const { installed_app_id, role_id, client, admin } = await installAppAndDna(
+      ADMIN_PORT
+    );
+    const createCloneCellParams: CreateCloneCellRequest = {
+      app_id: installed_app_id,
+      role_id,
+      modifiers: {
+        network_seed: "clone-0",
+      },
+    };
+    const cloneCell = await client.createCloneCell(createCloneCellParams);
+    await client.archiveCloneCell({
+      app_id: installed_app_id,
+      clone_cell_id: cloneCell.cell_id,
+    });
+
+    await admin.restoreCloneCell({
+      app_id: installed_app_id,
+      clone_cell_id: CloneId.fromRoleId(cloneCell.role_id).toString(),
+    });
+
+    const appInfo = await client.appInfo({ installed_app_id });
+    t.equal(
+      appInfo.cell_data.length,
+      2,
+      "restored clone cell is part of app info"
+    );
+    const params: CallZomeRequest = {
+      cap_secret: null,
+      cell_id: cloneCell.cell_id,
+      zome_name: TEST_ZOME_NAME,
+      fn_name: "foo",
+      provenance: fakeAgentPubKey(),
+      payload: null,
+    };
+    const resopnse = await client.callZome(params);
+    t.equal(resopnse, "foo", "restored clone cell can be called");
+  })
+);
+
+test(
+  "can delete archived clone cells of an app",
+  withConductor(ADMIN_PORT, async (t: Test) => {
+    const { installed_app_id, role_id, client, admin } = await installAppAndDna(
+      ADMIN_PORT
+    );
+    const createCloneCellParams: CreateCloneCellRequest = {
+      app_id: installed_app_id,
+      role_id,
+      modifiers: {
+        network_seed: "clone-0",
+      },
+    };
+    const cloneCell0 = await client.createCloneCell(createCloneCellParams);
+    createCloneCellParams.modifiers.network_seed = "clone-1";
+    const cloneCell1 = await client.createCloneCell(createCloneCellParams);
+    await client.archiveCloneCell({
+      app_id: installed_app_id,
+      clone_cell_id: cloneCell0.cell_id,
+    });
+    await client.archiveCloneCell({
+      app_id: installed_app_id,
+      clone_cell_id: cloneCell1.cell_id,
+    });
+
+    await admin.deleteArchivedCloneCells({ app_id: installed_app_id, role_id });
+
+    try {
+      await admin.restoreCloneCell({
+        app_id: installed_app_id,
+        clone_cell_id: cloneCell0.cell_id,
+      });
+      t.fail();
+    } catch (error) {
+      t.pass("deleted clone cell 0 cannot be restored");
+    }
+    try {
+      await admin.restoreCloneCell({
+        app_id: installed_app_id,
+        clone_cell_id: cloneCell1.cell_id,
+      });
+      t.fail();
+    } catch (error) {
+      t.pass("deleted clone cell 1 cannot be restored");
+    }
   })
 );